requirements.txt
viam-sdk
openmeteo-requests
requests-cache
retry-requests
numpy
panda